app/requirements.txt
Django==3.0.*
djangorestframework==3.11.*
django-nested-inline==0.4.*
django-url-filter==0.3.*
drf-nested-routers==0.*
gunicorn==20.0.*
Pillow==7.1.*
psycopg2==2.8.*
